+++ app/javascript/Components/instructor_table.jsx
@@ -4,6 +4,7 @@ import PropTypes from "prop-types";
 
 import ReactTable from "react-table";
 import {selectFilter} from "./Helpers/table_helpers";
+import {tableNoDataComponent} from "./table_no_data";
 
 class InstructorTable extends React.Component {
   constructor() {
@@ -100,6 +101,7 @@ class InstructorTable extends React.Component {
         ]}
         filterable
         loading={this.state.loading}
+        NoDataComponent={() => tableNoDataComponent(I18n.t("instructors.empty_table"))}
       />
     );
   }