+++ app/javascript/Components/table_no_data.jsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export const tableNoDataComponent = empty_table_message => {
+  return <p className="rt-no-data">{empty_table_message}</p>;
+};